+++ datalad/customremotes/tests/test_archives.py
@@ -8,6 +8,8 @@
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Tests for customremotes archives providing dl+archive URLs handling"""
 
+from datalad.tests.utils import skip_v6
+from datalad.tests.utils import skip_direct_mode
 from ..archives import ArchiveAnnexCustomRemote
 from ..base import AnnexExchangeProtocol
 from ...support.annexrepo import AnnexRepo
@@ -119,6 +121,8 @@ def check_basic_scenario(fn_archive, fn_extracted, direct, d, d2):
 @with_tree(
     tree={'a.tar.gz': {'d': {fn_inarchive_obscure: '123'}}}
 )
+@skip_direct_mode  #FIXME
+@skip_v6  #FIXME
 def test_annex_get_from_subdir(topdir):
     from datalad.api import add_archive_content
     annex = AnnexRepo(topdir, init=True)
@@ -151,6 +155,8 @@ def test_get_git_environ_adjusted():
     assert_equal(sys_env["PWD"], os.environ.get("PWD"))
 
 
+@skip_direct_mode  #FIXME
+@skip_v6  #FIXME
 def test_basic_scenario():
     yield check_basic_scenario, 'a.tar.gz', 'simple.txt', False
     if not on_windows: