+++ datalad/distribution/tests/test_install.py
@@ -9,6 +9,8 @@
 
 """
 
+from datalad.tests.utils import skip_v6
+from datalad.tests.utils import skip_direct_mode
 import logging
 import os
 
@@ -46,6 +48,7 @@ from datalad.tests.utils import assert_false
 from datalad.tests.utils import ok_file_has_content
 from datalad.tests.utils import assert_not_in
 from datalad.tests.utils import assert_raises
+from datalad.tests.utils import assert_is_instance
 from datalad.tests.utils import assert_result_count
 from datalad.tests.utils import assert_status
 from datalad.tests.utils import assert_in_results
@@ -220,6 +223,7 @@ def test_install_datasets_root(tdir):
 
 @with_testrepos('.*basic.*', flavors=['local-url', 'network', 'local'])
 @with_tempfile(mkdir=True)
+@skip_v6  #FIXME
 def test_install_simple_local(src, path):
     origin = Dataset(path)
 
@@ -259,6 +263,7 @@ def test_install_simple_local(src, path):
 
 @with_testrepos(flavors=['local-url', 'network', 'local'])
 @with_tempfile
+@skip_v6  #FIXME
 def test_install_dataset_from_just_source(url, path):
     with chpwd(path, mkdir=True):
         ds = install(source=url)
@@ -270,8 +275,25 @@ def test_install_dataset_from_just_source(url, path):
     assert_in('INFO.txt', ds.repo.get_indexed_files())
 
 
+@with_testrepos(flavors=['local'])
+@with_tempfile(mkdir=True)
+@skip_v6  #FIXME
+def test_install_dataset_from_instance(src, dst):
+    origin = Dataset(src)
+    clone = install(source=origin, path=dst)
+
+    assert_is_instance(clone, Dataset)
+    ok_startswith(clone.path, dst)
+    ok_(clone.is_installed())
+    ok_(GitRepo.is_valid_repo(clone.path))
+    ok_clean_git(clone.path, annex=None)
+    assert_in('INFO.txt', clone.repo.get_indexed_files())
+
+
 @with_testrepos(flavors=['network'])
 @with_tempfile
+@skip_v6  #FIXME
+@skip_v6  #FIXME
 def test_install_dataset_from_just_source_via_path(url, path):
     # for remote urls only, the source could be given to `path`
     # to allows for simplistic cmdline calls
@@ -311,6 +333,7 @@ def test_install_dataladri(src, topurl, path):
 @with_testrepos('submodule_annex', flavors=['local', 'local-url', 'network'])
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_v6  #FIXME
 def test_install_recursive(src, path_nr, path_r):
     # first install non-recursive:
     ds = install(path_nr, source=src, recursive=False)
@@ -354,6 +377,8 @@ def test_install_recursive(src, path_nr, path_r):
 
 @with_testrepos('submodule_annex', flavors=['local'])
 @with_tempfile(mkdir=True)
+@skip_v6  #FIXME
+@skip_v6  #FIXME
 def test_install_recursive_with_data(src, path):
 
     # now again; with data:
@@ -384,6 +409,7 @@ def test_install_recursive_with_data(src, path):
 # .git/config show a submodule url "file:///aaa/bbb%20b/..."
 # this is delivered by with_testrepos as the url to clone
 @with_tempfile
+@skip_direct_mode  #FIXME
 def test_install_into_dataset(source, top_path):
 
     ds = create(top_path)
@@ -425,6 +451,7 @@ def test_install_into_dataset(source, top_path):
 @skip_if_no_network
 @use_cassette('test_install_crcns')
 @with_tempfile
+@skip_direct_mode  #FIXME
 def test_failed_install_multiple(top_path):
     ds = create(top_path)
 
@@ -450,6 +477,7 @@ def test_failed_install_multiple(top_path):
 
 @with_testrepos('submodule_annex', flavors=['local', 'local-url', 'network'])
 @with_tempfile(mkdir=True)
+@skip_v6  #FIXME
 def test_install_known_subdataset(src, path):
 
     # get the superdataset:
@@ -482,6 +510,7 @@ def test_install_known_subdataset(src, path):
 @slow  # 46.3650s
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_implicit_install(src, dst):
 
     origin_top = create(src)
@@ -598,6 +627,8 @@ def test_reckless(path, top_path):
                            }
                  })
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
+@skip_v6  #FIXME
 def test_install_recursive_repeat(src, path):
     subsub_src = Dataset(opj(src, 'sub 1', 'subsub')).create(force=True)
     sub1_src = Dataset(opj(src, 'sub 1')).create(force=True)
@@ -711,6 +742,7 @@ def test_install_skip_failed_recursive(src, path):
                            }
                  })
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_install_noautoget_data(src, path):
     subsub_src = Dataset(opj(src, 'sub 1', 'subsub')).create(force=True)
     sub1_src = Dataset(opj(src, 'sub 1')).create(force=True)
@@ -741,6 +773,7 @@ def test_install_source_relpath(src, dest):
 @with_tempfile
 @with_tempfile
 @with_tempfile
+@skip_direct_mode  #FIXME
 def test_install_consistent_state(src, dest, dest2, dest3):
     # if we install a dataset, where sub-dataset "went ahead" in that branch,
     # while super-dataset was not yet updated (e.g. we installed super before)