+++ datalad/distribution/tests/test_update.py
@@ -9,6 +9,8 @@
 
 """
 
+from datalad.tests.utils import skip_v6
+from datalad.tests.utils import skip_direct_mode
 import os
 from os.path import join as opj, exists
 from ..dataset import Dataset
@@ -34,6 +36,8 @@ from datalad.tests.utils import assert_in_results
 @with_testrepos('submodule_annex', flavors=['local'])  #TODO: Use all repos after fixing them
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
+@skip_v6  #FIXME
 def test_update_simple(origin, src_path, dst_path):
 
     # prepare src
@@ -123,6 +127,7 @@ def test_update_git_smoke(src_path, dst_path):
 @with_testrepos('.*annex.*', flavors=['clone'])
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_update_fetch_all(src, remote_1, remote_2):
     rmt1 = AnnexRepo.clone(src, remote_1)
     rmt2 = AnnexRepo.clone(src, remote_2)
@@ -181,6 +186,7 @@ def test_update_fetch_all(src, remote_1, remote_2):
 
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_newthings_coming_down(originpath, destpath):
     origin = GitRepo(originpath, create=True)
     create_tree(originpath, {'load.dat': 'heavy'})
@@ -237,6 +243,7 @@ def test_newthings_coming_down(originpath, destpath):
 
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_update_volatile_subds(originpath, destpath):
     origin = Dataset(originpath).create()
     ds = install(
@@ -285,6 +292,7 @@ def test_update_volatile_subds(originpath, destpath):
 
 @with_tempfile(mkdir=True)
 @with_tempfile(mkdir=True)
+@skip_direct_mode  #FIXME
 def test_reobtain_data(originpath, destpath):
     origin = Dataset(originpath).create()
     ds = install(