+++ datalad/interface/diff.py
@@ -15,7 +15,9 @@ import logging
 import stat
 from os.path import join as opj
 from os.path import curdir
+from os.path import isdir
 from os.path import relpath
+from os.path import normpath
 
 
 from datalad.interface.annotate_paths import AnnotatePaths
@@ -27,7 +29,6 @@ from datalad.support.constraints import EnsureNone
 from datalad.support.constraints import EnsureStr
 from datalad.support.constraints import EnsureChoice
 from datalad.support.exceptions import CommandError
-from datalad.support.exceptions import InsufficientArgumentsError
 from datalad.support.param import Parameter
 from datalad.interface.common_opts import recursion_flag
 from datalad.interface.common_opts import recursion_limit
@@ -35,6 +36,11 @@ from datalad.cmd import GitRunner
 
 from datalad.distribution.dataset import EnsureDataset
 from datalad.distribution.dataset import datasetmethod
+
+from datalad.utils import with_pathsep as _with_sep
+
+from datalad.consts import PRE_INIT_COMMIT_SHA
+
 
 lgr = logging.getLogger('datalad.interface.diff')
 
@@ -65,10 +71,59 @@ def _translate_type(mode, ap, prop):
     elif mode == stat.S_IFDIR | stat.S_IFLNK:
         ap[prop] = 'dataset'
     elif stat.S_ISDIR(mode):
-        # not sure if this can happen at all
+        # not sure if this can happen at all, at least not in the tests...
         ap[prop] = 'directory'
     else:
         ap[prop] = 'file'
+
+
+def _get_untracked_content(dspath, report_untracked, paths=None):
+    cmd = ['git', '--work-tree=.', 'status', '--porcelain',
+           # file names NULL terminated
+           '-z',
+           # we never want to touch submodules, they cannot be untracked
+           '--ignore-submodules=all',
+           # fully untracked dirs as such, the rest as files
+           '--untracked={}'.format(report_untracked)]
+    try:
+        stdout, stderr = GitRunner(cwd=dspath).run(
+            cmd,
+            log_stderr=True,
+            log_stdout=True,
+            log_online=False,
+            expect_stderr=False,
+            shell=False,
+            expect_fail=True)
+    except CommandError as e:
+        # TODO should we catch any and handle them in here?
+        raise e
+
+    if paths:
+        paths = [r['path'] for r in paths]
+        if len(paths) == 1 and paths[0] == dspath:
+            # nothing to filter
+            paths = None
+
+    for line in stdout.split('\0'):
+        if not line:
+            continue
+        if not line.startswith('?? '):
+            # nothing untracked, ignore, task of `diff`
+            continue
+        apath = opj(
+            dspath,
+            # strip state marker
+            line[3:])
+        norm_apath = normpath(apath)
+        if paths and not any([norm_apath == p or apath.startswith(_with_sep(p)) for p in paths]):
+            # we got a whitelist for paths, don't report any other
+            continue
+        ap = dict(
+            path=norm_apath,
+            parentds=dspath,
+            state='untracked',
+            type='directory' if isdir(apath) else 'file')
+        yield ap
 
 
 def _parse_git_diff(dspath, diff_thingie=None, paths=None,
@@ -141,9 +196,34 @@ def _parse_git_diff(dspath, diff_thingie=None, paths=None,
 
 @build_doc
 class Diff(Interface):
-    """Report changes of dataset component between revisions.
+    """Report changes of dataset components.
+
+    Reports can be generated for changes between recorded revisions, or
+    between a revision and the state of a dataset's work tree.
+
+    Unlike 'git diff', this command also reports untracked content when
+    comparing a revision to the state of the work tree. Such content is
+    marked with the property `state='untracked'` in the command results.
+
+    The following types of changes are distinguished and reported via the
+    `state` result property:
+
+    - added
+    - copied
+    - deleted
+    - modified
+    - renamed
+    - typechange
+    - unmerged
+    - untracked
+
+    Whenever applicable, source and/or destination revisions are reported
+    to indicate when exactly within the requested revision range a particular
+    component changed its status.
+
+    Optionally, the reported changes can be limited to a subset of paths
+    within a dataset.
     """
-    # TODO describe properties that are reported
 
     # make the custom renderer the default one, as the global default renderer
     # does not yield meaningful output for this command
@@ -194,6 +274,14 @@ class Diff(Interface):
             execution will still report other changes in any existing
             subdataset, only the subdataset record in a parent dataset
             is not  evaluated."""),
+        report_untracked=Parameter(
+            args=('--report-untracked',),
+            constraints=EnsureChoice('no', 'normal', 'all'),
+            doc="""If and how untracked content is reported when comparing
+            a revision to the state of the work tree. 'no': no untracked files
+            are reported; 'normal': untracked files and entire untracked
+            directories are reported as such; 'all': report individual files
+            even in fully untracked directories."""),
         recursive=recursion_flag,
         recursion_limit=recursion_limit)
 
@@ -203,20 +291,20 @@ class Diff(Interface):
     def __call__(
             path=None,
             dataset=None,
-            revision='HEAD',
+            revision=None,
             staged=False,
             ignore_subdatasets='none',
+            report_untracked='normal',
             recursive=False,
             recursion_limit=None):
         if not dataset and not path:
             # act on the whole dataset if nothing else was specified
             dataset = curdir
         refds_path = Interface.get_refds_path(dataset)
-        if not (refds_path or path):
-            raise InsufficientArgumentsError(
-                "Neither dataset nor target path(s) provided")
 
         to_process = []
+        # tracked what commit ranges we want to diff per dataset
+        ds_diffies = {}
         for ap in AnnotatePaths.__call__(
                 path=path,
                 dataset=refds_path,
@@ -226,6 +314,8 @@ class Diff(Interface):
                 # unavailable is OK, because we might query for a deleted file
                 unavailable_path_status='',
                 nondataset_path_status='impossible',
+                # must not use `modified`, infinite loop otherwise
+                modified=None,
                 return_type='generator',
                 on_failure='ignore'):
             if ap.get('status', None):
@@ -234,6 +324,11 @@ class Diff(Interface):
                 continue
             if ap.get('type', None) == 'dataset':
                 ap['process_content'] = True
+            if ap.get('raw_input', False) or ap['path'] == refds_path:
+                # prepopulate the revision specs for all input paths
+                ds_diffies[ap['path']
+                           if ap.get('type', None) == 'dataset'
+                           else ap['parentds']] = revision
             to_process.append(ap)
 
         # sort into datasets
@@ -245,16 +340,59 @@ class Diff(Interface):
         assert(not completed)
 
         for ds_path in sorted(content_by_ds.keys()):
+            if ds_path not in ds_diffies:
+                # we don't know how to diff
+                # this was not neither an input path, not did we see it
+                # when diffing its parent
+                continue
+            content_paths = content_by_ds[ds_path]
+            revision = ds_diffies[ds_path]
             for r in _parse_git_diff(
                     ds_path,
-                    diff_thingie=revision,
-                    paths=content_by_ds[ds_path],
+                    diff_thingie=ds_diffies[ds_path],
+                    paths=content_paths,
                     ignore_submodules=ignore_subdatasets,
                     staged=staged):
                 r.update(dict(
                     action='diff',
-                    refds=refds_path),
-                    logger=lgr)
+                    logger=lgr))
+                if refds_path:
+                    r['refds'] = refds_path
+                if 'status' not in r:
+                    r['status'] = 'ok'
+                if r.get('type', None) == 'dataset':
+                    # this is a subdataset report
+                    # we need to use the reported commit range to properly adjust the
+                    # query once we hit that subdataset
+                    from_rev = r.get('revision_src', '')
+                    to_rev = r.get('revision', '')
+                    subrev = '{}..{}'.format(
+                        from_rev if from_rev else PRE_INIT_COMMIT_SHA,
+                        to_rev if to_rev else '',
+                    )
+                    if from_rev and from_rev == to_rev:
+                        # this is a special case, where subdataset reported changes without
+                        # a change in state/commit -- this is code for uncommited changes
+                        # in the subdataset (including staged ones). In such a case, we
+                        # must not provide a diff range, but only the source commit we want
+                        # to diff against
+                        # XXX if this is changed, likely the same logic in annotate_paths needs
+                        # changing too!
+                        subrev = from_rev
+                    ds_diffies[r['path']] = subrev
+                yield r
+            if (revision and '..' in revision) or report_untracked == 'no':
+                # don't look for untracked content, we got a revision range
+                continue
+            for r in _get_untracked_content(
+                    ds_path,
+                    report_untracked,
+                    paths=content_paths):
+                r.update(dict(
+                    action='diff',
+                    logger=lgr))
+                if refds_path:
+                    r['refds'] = refds_path
                 if 'status' not in r:
                     r['status'] = 'ok'
                 yield r
@@ -266,9 +404,9 @@ class Diff(Interface):
             # logging reported already
             return
         path = relpath(res['path'], start=res['refds']) \
-            if 'refds' in res else res['path']
+            if res.get('refds', None) else res['path']
         type_ = res.get('type', res.get('type_src', ''))
-        max_len = len('typechange(dataset)')
+        max_len = len('untracked(directory)')
         state_msg = '{}{}'.format(
             res['state'],
             '({})'.format(type_ if type_ else ''))