+++ datalad/interface/save.py
@@ -20,10 +20,12 @@ from os.path import lexists
 
 
 from datalad.utils import unique
+from datalad.utils import swallow_logs
 from datalad.support.annexrepo import AnnexRepo
 from datalad.support.constraints import EnsureStr
 from datalad.support.constraints import EnsureNone
 from datalad.support.param import Parameter
+from datalad.support.exceptions import CommandError
 from datalad.distribution.dataset import Dataset
 from datalad.distribution.dataset import EnsureDataset
 from datalad.distribution.dataset import datasetmethod
@@ -46,8 +48,7 @@ lgr = logging.getLogger('datalad.interface.save')
 def save_dataset(
         ds,
         paths,
-        message=None,
-        version_tag=None):
+        message=None):
     """Save changes in a single dataset.
 
     Parameters
@@ -58,8 +59,6 @@ def save_dataset(
       Annotated paths to dataset components to be saved.
     message: str, optional
       (Commit) message to be attached to the saved state.
-    version_tag : str, optional
-      Tag to be assigned to the saved state.
 
     Returns
     -------
@@ -129,11 +128,6 @@ def save_dataset(
         _datalad_msg=_datalad_msg,
         careless=True)
 
-    # MIH: let's tag even if there was nothing commit. I'd forget this
-    # option too often...
-    if version_tag:
-        ds.repo.tag(version_tag)
-
     current_hexsha = ds.repo.get_hexsha()
     _was_modified = current_hexsha != orig_hexsha
     return current_hexsha if _was_modified else None
@@ -163,11 +157,11 @@ class Save(Interface):
             doc=""""specify the dataset to save. If a dataset is given, but
             no `files`, the entire dataset will be saved.""",
             constraints=EnsureDataset() | EnsureNone()),
-        files=Parameter(
-            args=("files",),
-            metavar='FILES',
-            doc="""list of files to consider. If given, only changes made
-            to those files are recorded in the new state.""",
+        path=Parameter(
+            args=("path",),
+            metavar='PATH',
+            doc="""path/name of the dataset component to save. If given, only
+            changes made to those components are recorded in the new state.""",
             nargs='*',
             constraints=EnsureStr() | EnsureNone()),
         message=save_message_opt,
@@ -194,8 +188,7 @@ class Save(Interface):
     @staticmethod
     @datasetmethod(name='save')
     @eval_results
-    # TODO files -> path
-    def __call__(message=None, files=None, dataset=None,
+    def __call__(message=None, path=None, dataset=None,
                  all_updated=True, all_changes=None, version_tag=None,
                  recursive=False, recursion_limit=None, super_datasets=False
                  ):
@@ -206,15 +199,16 @@ class Save(Interface):
                 version="0.5.0",
                 msg="RF: all_changes option passed to the save"
             )
-        if not dataset and not files:
+        if not dataset and not path:
             # we got nothing at all -> save what is staged in the repo in "this" directory?
             # we verify that there is an actual repo next
             dataset = abspath(curdir)
         refds_path = Interface.get_refds_path(dataset)
 
         to_process = []
+        got_nothing = True
         for ap in AnnotatePaths.__call__(
-                path=files,
+                path=path,
                 dataset=refds_path,
                 recursive=recursive,
                 recursion_limit=recursion_limit,
@@ -222,8 +216,14 @@ class Save(Interface):
                 unavailable_path_status='impossible',
                 unavailable_path_msg="path does not exist: %s",
                 nondataset_path_status='impossible',
+                modified='HEAD' if not path and recursive else None,
                 return_type='generator',
                 on_failure='ignore'):
+            if ap.get('state', None) == 'untracked' and not ap.get('raw_input', False):
+                # this path was found untracked, but not explicitly given to save
+                # we will silently ignore this
+                continue
+            got_nothing = False
             # next check should not be done during annotation, as it is possibly expensive
             # and not generally useful
             if ap.get('status', None) == 'impossible' and \
@@ -249,6 +249,18 @@ class Save(Interface):
                 ap['process_content'] = True
                 ap['process_updated_only'] = all_updated
             to_process.append(ap)
+
+        if got_nothing and recursive and refds_path:
+            # path annotation yielded nothing, most likely cause is that nothing
+            # was found modified, we need to say something about the reference
+            # dataset
+            yield get_status_dict(
+                'save',
+                status='notneeded',
+                path=refds_path,
+                type='dataset',
+                logger=lgr)
+            return
 
         if not to_process:
             # nothing left to do, potentially all errored before
@@ -313,7 +325,10 @@ class Save(Interface):
                     path=parentds,
                     type='dataset',
                     # make sure we save content of superds later on
-                    process_content=True))
+                    process_content=True,
+                    # but not do nasty things, like adding untracked content
+                    # just because we discovered this dataset
+                    process_updated_only=True))
 
         # now re-annotate all paths, this will be fast for already annotated ones
         # and will amend the annotation for others, deduplication happens here too
@@ -350,13 +365,35 @@ class Save(Interface):
             saved_state = save_dataset(
                 ds,
                 content_by_ds[dspath],
-                message=message,
-                version_tag=version_tag)
-            if saved_state:
-                res['status'] = 'ok'
+                message=message)
+            res['status'] = 'ok' if saved_state else 'notneeded'
+            # MIH: let's tag even if there was nothing commit. I'd forget this
+            # option too often...
+            if version_tag:
+                try:
+                    # again cannot help but force-silence low-level code, because
+                    # it screams like a made man instead of allowing top-level
+                    # code an orderly error report
+                    with swallow_logs():
+                        ds.repo.tag(version_tag)
+                    # even if we haven't saved anything
+                    res['status'] = 'ok'
+                    yield res
+                except CommandError as e:
+                    if saved_state:
+                        # first we yield the result for the actual save
+                        yield res
+                    # and now complain that tagging didn't work
+                    yield get_status_dict(
+                        'save',
+                        ds=ds,
+                        logger=lgr,
+                        status='error',
+                        message=(
+                            'cannot tag this version: %s',
+                            e.stderr.strip()))
             else:
-                res['status'] = 'notneeded'
-            yield res
+                yield res
 
     @staticmethod
     def custom_result_renderer(res, **kwargs):