+++ datalad/support/annexrepo.py
@@ -78,6 +78,8 @@ from .exceptions import RemoteNotAvailableError
 from .exceptions import OutdatedExternalDependency
 from .exceptions import MissingExternalDependency
 from .exceptions import IncompleteResultsError
+from .exceptions import AccessDeniedError
+from .exceptions import AccessFailedError
 
 lgr = logging.getLogger('datalad.annex')
 
@@ -268,19 +270,19 @@ class AnnexRepo(GitRepo, RepoInterface):
         # Figure out, whether we want to allow for a list here or what to
         # do, if there is sth in that setting already
         if persistent:
-            git_attributes_file = _path_(self.path, '.gitattributes')
-            git_attributes = ''
-            if exists(git_attributes_file):
-                with open(git_attributes_file) as f:
-                    git_attributes = f.read()
-            if ' annex.backend=' in git_attributes:
+            # could be set in .gitattributes or $GIT_DIR/info/attributes
+            if 'annex.backend' in self.get_git_attributes():
                 lgr.debug(
-                    "Not (re)setting backend since seems already set in %s"
-                    % git_attributes_file
+                    "Not (re)setting backend since seems already set in git attributes"
                 )
             else:
                 lgr.debug("Setting annex backend to %s (persistently)", backend)
                 self.config.set('annex.backends', backend, where='local')
+                git_attributes_file = _path_(self.path, '.gitattributes')
+                git_attributes = ''
+                if exists(git_attributes_file):
+                    with open(git_attributes_file) as f:
+                        git_attributes = f.read()
                 with open(git_attributes_file, 'a') as f:
                     if git_attributes and not git_attributes.endswith(os.linesep):
                         f.write(os.linesep)
@@ -919,6 +921,41 @@ class AnnexRepo(GitRepo, RepoInterface):
         else:
             return remotes
 
+    def get_special_remotes(self):
+        """Get info about all known (not just enabled) special remotes.
+
+        Returns
+        -------
+        dict
+          Keys are special remore UUIDs, values are dicts with arguments
+          for `git-annex enableremote`. This includes at least the 'type'
+          and 'name' of a special remote. Each type of special remote
+          may require addition arguments that will be available in the
+          respective dictionary.
+        """
+        try:
+            stdout, stderr = self._git_custom_command(
+                None, ['git', 'cat-file', 'blob', 'git-annex:remote.log'],
+                expect_fail=True)
+        except CommandError as e:
+            if 'Not a valid object name git-annex:remote.log' in e.stderr:
+                # no special remotes configures
+                return {}
+            else:
+                # some unforseen error
+                raise e
+        argspec = re.compile(r'^([^=]*)=(.*)$')
+        srs = {}
+        for line in stdout.splitlines():
+            # be precise and split by spaces
+            fields = line.split(' ')
+            # special remote UUID
+            sr_id = fields[0]
+            # the rest are config args for enableremote
+            sr_info = dict(argspec.match(arg).groups()[:2] for arg in fields[1:])
+            srs[sr_id] = sr_info
+        return srs
+
     def __repr__(self):
         return "<AnnexRepo path=%s (%s)>" % (self.path, type(self))
 
@@ -1415,7 +1452,8 @@ class AnnexRepo(GitRepo, RepoInterface):
 
             # if None -- leave it to annex to decide
             if git is not None:
-                if self.config.getint("annex", "version") == 6:
+                if 'annex.version' in self.config and \
+                        self.config.getint("annex", "version") == 6:
                     # Note: For now ugly workaround to prevent unexpected
                     # outcome when adding to git. See:
                     # <http://git-annex.branchable.com/bugs/mysterious_dependency_of_git_annex_status_output_of_the_added_file/>
@@ -1797,7 +1835,7 @@ class AnnexRepo(GitRepo, RepoInterface):
         self._run_annex_command('initremote', annex_options=[name] + options)
         self.config.reload()
 
-    def enable_remote(self, name):
+    def enable_remote(self, name, env=None):
         """Enables use of an existing special remote
 
         Parameters
@@ -1806,7 +1844,20 @@ class AnnexRepo(GitRepo, RepoInterface):
             name, the special remote was created with
         """
 
-        self._run_annex_command('enableremote', annex_options=[name])
+        try:
+            self._run_annex_command(
+                'enableremote',
+                annex_options=[name],
+                expect_fail=True,
+                log_stderr=True,
+                env=env)
+        except CommandError as e:
+            if re.match(r'.*StatusCodeException.*statusCode = 401', e.stderr):
+                raise AccessDeniedError(e.stderr)
+            elif 'FailedConnectionException' in e.stderr:
+                raise AccessFailedError(e.stderr)
+            else:
+                raise e
         self.config.reload()
 
     def merge_annex(self, remote=None):